Move ZBar.java to com.sparrowwallet.sparrow.control package
